+++ burpui/misc/backend/burp1.py
@@ -1013,7 +1013,8 @@ class Burp(BUIbackend):
             return []
         return self.parser.read_server_conf(conf)
 
-    def store_conf_cli(self, data, client=None, conf=None, template=False, agent=None):
+    def store_conf_cli(self, data, client=None, conf=None, template=False,
+                       statictemplate=False, content='', agent=None):
         """See :func:`burpui.misc.backend.interface.BUIbackend.store_conf_cli`"""
         if not self.parser:
             return []
@@ -1021,7 +1022,7 @@ class Burp(BUIbackend):
             conf = unquote(conf)
         except:
             pass
-        return self.parser.store_client_conf(data, client, conf, template)
+        return self.parser.store_client_conf(data, client, conf, template, statictemplate, content)
 
     def store_conf_srv(self, data, conf=None, agent=None):
         """See :func:`burpui.misc.backend.interface.BUIbackend.store_conf_srv`"""